refactor: address review comments, move vnish related functions to crate
