+++ asic-rs-firmwares/vnish/Cargo.toml
@@ -13,6 +13,7 @@ asic-rs-core.workspace = true
 asic-rs-makes-antminer.workspace = true
 
 async-trait.workspace = true
+chrono.workspace = true
 reqwest.workspace = true
 once_cell.workspace = true
 serde_json.workspace = true
