+++ asic-rs-firmwares/vnish/src/lib.rs
@@ -1,2 +1,3 @@
 mod backends;
 pub mod firmware;
+mod timezone;
